% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Linear regression.}
\description{
Performs linear regression using QR decomposition.
See \url{http://staff.www.ltu.se/~jove/courses/c0002m/least squares.pdf}
for details.
}
\section{Fields}{

\describe{
\item{\code{formula}}{formula Linear regression formula.}

\item{\code{data}}{data.frame Input data.}

\item{\code{coefficients}}{numeric Coefficients.}

\item{\code{residuals}}{matrix Matrix of residuals.}

\item{\code{fitted.values}}{matrix Matrix of fitted values.}

\item{\code{df}}{numeric Degrees of freedom.}

\item{\code{sigma}}{numeric Estimated residual standard error.}

\item{\code{X}}{matrix Model matrix.}

\item{\code{y}}{numeric Response vector.}

\item{\code{qr}}{ANY Results of the QR decomposition.}
}}


\examples{
linreg_mod <- linreg$new(formula = Petal.Length ~ Species, data=iris)
linreg_mod$plot()
linreg_mod$print()
linreg_mod$resid()
linreg_mod$pred()
linreg_mod$coef()
linreg_mod$summary()
}
