% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{predict.ridgereg}
\alias{predict.ridgereg}
\title{Generates predictions from a fitted ridgereg model.}
\usage{
\method{predict}{ridgereg}(object, newdata)
}
\arguments{
\item{object}{A fitted object of class 'ridgereg'.}

\item{newdata}{A data frame for prediction. If missing, fitted values are returned.}
}
\description{
Generates predictions from a fitted ridgereg model.
}
