% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{ridgereg}
\alias{ridgereg}
\title{Ridge regression}
\usage{
ridgereg(formula, data, lambda)
}
\arguments{
\item{formula}{formula Linear regression formula.}

\item{data}{data.frame Input data.}

\item{lambda}{numeric A non-negative regularization parameter}
}
\value{
ridgereg object containing the results.
}
\description{
Performs ridge regression using a formula interface.See
\url{http://math.stackexchange.com/questions/299481/qr-factorization-for-ridge-regression}
for details.
}
\examples{
data(iris)
ridge_model <- ridgereg(Sepal.Length ~ Petal.Length + Petal.Width, data = iris, lambda = 0.5)
print(ridge_model)
}
