% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizualizeAirportDelay.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{Visualize Mean Airport Delays}
\usage{
visualize_airport_delays()
}
\value{
A ggplot object showing mean delays by airport location.
}
\description{
Creates a scatter plot showing the mean departure delay for different airports
across the United States, using longitude and latitude from the `airports` dataset.
}
\details{
This function uses data from the `nycflights13` package, combines the `flights`
and `airports` datasets, and visualizes average delays with `ggplot2`.
}
\examples{
\dontrun{
visualize_airport_delays()
}

}
