% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linregclassRC.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Title}
\description{
Title
}
\section{Fields}{

\describe{
\item{\code{coefficients}}{matrix.}

\item{\code{fittedvalues}}{matrix.}

\item{\code{residuals}}{matrix.}

\item{\code{df}}{numeric.}

\item{\code{residualvariance}}{numeric.}

\item{\code{varianceregcoe}}{matrix.}

\item{\code{tvalues}}{matrix.}

\item{\code{pvalue}}{matrix.}

\item{\code{formula}}{character.}

\item{\code{data}}{character.}

\item{\code{star}}{character.}
}}


