% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linregclassRC.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{A Reference Class to represent a linear model.}
\description{
A Reference Class to represent a linear model.
}
\section{Fields}{

\describe{
\item{\code{coefficients}}{Unknown parameters that describe the relationship between X and Y.}

\item{\code{fittedvalues}}{The predicted values of Y.}

\item{\code{residuals}}{difference between fitted values and Y.}

\item{\code{df}}{Difference between n the number of observations and p is the number of parameters in the model.}

\item{\code{residualvariance}}{Variance of residuals.}

\item{\code{varianceregcoe}}{Variance of Coefficients.}

\item{\code{tvalues}}{t-values of Coefficients.}

\item{\code{pvalue}}{P-values of Coefficients.}

\item{\code{formula}}{A formula object like x ~ y.}

\item{\code{data}}{A data frame which contains the formula variables.}

\item{\code{star}}{character.}
}}


