% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{A function for the multiple regression modEL}
\arguments{
\item{formula}{The formula argument should take a formula object. It includes the formula given from the user so that the function will do the calculations for the regression.}

\item{data}{It must be a data frame and it includes the data for the regression.}

\item{coefficients}{It must be a matrix. A matrix to store the regression coefficients.}

\item{residuals}{It must be an array. It is used to store the residuals.}

\item{predicted}{It must be a matrix. It is used to store the predicted values.}

\item{degrees_of_freedom}{The type should be numeric. It is used to store the degrees of freedom.}

\item{residual_variance}{A numeric value for the residual variance.}

\item{variance_of_coefficients}{It must be a matrix. It is used to store the variance of the coefficients.}

\item{t_values}{It must be a matrix to store the t values.}

\item{p_values}{It must be a matrix to store the p values.}
}
\value{
The function returns an object of class linreg as an RC class.
}
\description{
The function receives 2 arguments formula and data and uses linear algebra and returns of class linreg as an RC class.
A reference class called "linreg" is defined using "setRefClass". It is used to perform a linear regression analysis on a data set providing the summary of the results.The class and the methods are designed to make the linear regression model more user friendly and organised.

The fields list is created to define the fields that an object of the linreg class will have.

The methods list is created to define the methods (functions) that can be applied to objects of the linreg class.These methods assist to perform various operations on the linear regression model.

Initialize: Constructor method for creating a linreg object. It takes a formula and data as arguments, stores them, and then calls fit_model to perform the regression.
The fit model uses the provided formula and data, calculating coefficients, residuals, and various statistics.
The resid returns the residuals as a vector.
The pred returns the predicted values.
The coef returns the coefficients with their names.
The summary prints a summary of the linear regression model, including coefficients, standard errors, t-values, p-values, residual standard error, and variance of coefficients.
The show displays basic information about the linreg object, including the formula, data, and coefficients.
The plot generates and displays two diagnostic plots for the linear regression model: "Residuals vs Fitted" and "Scale-Location".
}

